WIP: end-to-end test (almost)
This test actually fetches data from the real API, and it passes (as
long as the API is running.) I can't keep this test at present because
it would need too much setting up. Also, to be truly end-to-end, the
test would need to have control of the server's disk, create a file on
it, and then check this file is one of the options on the page.