Move container cleanup step
There should be no interim commands between taking down the system and
running it again because they increase downtime and may even fail,
blocking deployment while the system is down. This moves all cleanup
commands to the end.

It might fail and then it will block deployment. Also it shouldn't run
at this point because it increases downtime even if it doesn't fail.