WIP prevent fetching on every click
This allows the app to use cache data if it isn't more than 5 minutes
old. The test is failing though. Before adding the staleTime parameter
it works as expected - the mock function is called 3 times. After adding
the staleTime parameter, the mock function is not being called, but the
assertions before it are passing. I need to look for any missing test
cleanup that might be causing this.